+++ src/test/java/org/plumelib/util/CollectionsPlumeTest.java
@@ -20,6 +20,7 @@ import java.util.Random;
 import java.util.Set;
 import java.util.SortedSet;
 import java.util.TreeSet;
+import java.util.function.Predicate;
 import org.checkerframework.checker.index.qual.IndexFor;
 import org.checkerframework.checker.lock.qual.GuardSatisfied;
 import org.junit.jupiter.api.Test;
@@ -73,11 +74,11 @@ public final class CollectionsPlumeTest {
     // Could make these lists immutable to avoid accidental modification in the tests.
   }
 
-  private static class OddFilter implements Filter<Integer> {
-    public OddFilter() {}
+  private static class OddPredicate implements Predicate<Integer> {
+    public OddPredicate() {}
 
     @Override
-    public boolean accept(Integer i) {
+    public boolean test(Integer i) {
       return i.intValue() % 2 != 0;
     }
   }
@@ -197,7 +198,7 @@ public final class CollectionsPlumeTest {
     assertEquals(
         iota10Odd,
         toArrayList(
-            new CollectionsPlume.FilteredIterator<Integer>(iota10.iterator(), new OddFilter())));
+            new CollectionsPlume.FilteredIterator<Integer>(iota10.iterator(), new OddPredicate())));
   }
 
   @Test
@@ -423,6 +424,30 @@ public final class CollectionsPlumeTest {
   }
 
   @Test
+  public void testAnyMatch() {
+    List<Integer> iota = Arrays.asList(new Integer[] {0, 1, 2, 3, 4, 5, 6, 7, 8, 9});
+    assertTrue(CollectionsPlume.anyMatch(iota, i -> i > -5));
+    assertTrue(CollectionsPlume.anyMatch(iota, i -> i > 5));
+    assertFalse(CollectionsPlume.anyMatch(iota, i -> i > 15));
+  }
+
+  @Test
+  public void testAllMatch() {
+    List<Integer> iota = Arrays.asList(new Integer[] {0, 1, 2, 3, 4, 5, 6, 7, 8, 9});
+    assertTrue(CollectionsPlume.allMatch(iota, i -> i > -5));
+    assertFalse(CollectionsPlume.allMatch(iota, i -> i > 5));
+    assertFalse(CollectionsPlume.allMatch(iota, i -> i > 15));
+  }
+
+  @Test
+  public void testNoneMatch() {
+    List<Integer> iota = Arrays.asList(new Integer[] {0, 1, 2, 3, 4, 5, 6, 7, 8, 9});
+    assertFalse(CollectionsPlume.noneMatch(iota, i -> i > -5));
+    assertFalse(CollectionsPlume.noneMatch(iota, i -> i > 5));
+    assertTrue(CollectionsPlume.noneMatch(iota, i -> i > 15));
+  }
+
+  @Test
   public void testListFilter() {
     List<Integer> in = Arrays.asList(new Integer[] {1, 2, 3, 4, 5});
     List<Integer> odd = Arrays.asList(new Integer[] {1, 3, 5});
@@ -567,6 +592,27 @@ public final class CollectionsPlumeTest {
       List<Integer> expected = l_101_103;
       assertEquals(expected, replaced);
     }
+  }
+
+  @Test
+  public void testIsSubsequenceMaybeNonContiguous() {
+    List<Integer> iota = Arrays.asList(new Integer[] {0, 1, 2, 3, 4, 5, 6, 7, 8, 9});
+    List<Integer> iota11 = Arrays.asList(new Integer[] {0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10});
+    List<Integer> empty = Arrays.asList(new Integer[] {});
+
+    assertTrue(CollectionsPlume.isSubsequenceMaybeNonContiguous(empty, empty));
+    assertTrue(CollectionsPlume.isSubsequenceMaybeNonContiguous(iota, empty));
+    assertFalse(CollectionsPlume.isSubsequenceMaybeNonContiguous(empty, iota));
+
+    assertTrue(CollectionsPlume.isSubsequenceMaybeNonContiguous(iota, Arrays.asList(0)));
+    assertTrue(CollectionsPlume.isSubsequenceMaybeNonContiguous(iota, Arrays.asList(0, 1, 2)));
+    assertTrue(CollectionsPlume.isSubsequenceMaybeNonContiguous(iota, Arrays.asList(9)));
+    assertTrue(CollectionsPlume.isSubsequenceMaybeNonContiguous(iota, Arrays.asList(7, 8, 9)));
+    assertTrue(CollectionsPlume.isSubsequenceMaybeNonContiguous(iota, Arrays.asList(5)));
+    assertTrue(CollectionsPlume.isSubsequenceMaybeNonContiguous(iota, Arrays.asList(4, 5, 6)));
+    assertTrue(CollectionsPlume.isSubsequenceMaybeNonContiguous(iota, Arrays.asList(2, 4, 6, 8)));
+    assertFalse(CollectionsPlume.isSubsequenceMaybeNonContiguous(iota, iota11));
+    assertTrue(CollectionsPlume.isSubsequenceMaybeNonContiguous(iota11, iota));
   }
 
   @Test